requirements.txt
Django==5.0.2
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
drf-spectacular==0.27.1
pillow==10.2.0
psycopg2==2.9.9
python-dotenv==1.0.1
gunicorn==20.0.4
uWSGI>=2.0.19.1,<2.1
